+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+8.24.3
+------
+* fix: create the `raw_ephys_data` folder even if there are no probes (when running behavior sessions on ephys rig)
+* move some Qt related code to `iblqt` repository
+
 8.24.2
 ------
 * make Frame2TTL validation more robust